+++ guestbook/models/__init__.py
@@ -1,0 +1,5 @@
+from _greeting import DEFAULT_GUESTBOOK_NAME, Greeting
+
+# def guestbook_key(guestbook_name=DEFAULT_GUESTBOOK_NAME):
+# 	'''Constructs a Datastore key for a Guestbook entity with guestbook_name.'''
+# 	return ndb.Key('Guestbook', guestbook_name)
