+++ guestbook/services/__init__.py
@@ -1,0 +1,1 @@
+from send_mail import add_queue_send_mail
