+++ guestbook/views/__init__.py
@@ -1,0 +1,2 @@
+from main_page import IndexView
+from sign_page import SignView
